+++ NEWS.md
@@ -1,10 +1,17 @@
+# usincometaxes 0.5.0
+
+### Major Changes
+
+- Add `interface` parameter to `taxsim_calculate_taxes()`, which allows users to select the methode of interfacing with TAXSIM.
+- Add 'wasm' option to `interface` parameter. The functionality incorporates Aman Gupta Karmani's  [JS / WebAssembly tooling](https://github.com/tmm1/taxsim.js) into the package.  Therefore, tax calculations can be conducted locally without send and retrieving data from the TAXSIM servers. (@thomascwells, [PR #11](https://github.com/shanejorr/usincometaxes/pull/11))
+- Add 'http' option to `interface` parameter. This option uses curl to send and retrieve data from TAXSIM via https. (@thomascwells, [PR #9](https://github.com/shanejorr/usincometaxes/pull/9))
+
 # usincometaxes 0.4.0
 
 ### Minor Changes
 
 - For ssh, create known_hosts file in temporary directory instead of .ssh.known_hosts. This is needed so that the package does not write to files outside the tempdir, in violation of CRAN requirements.
 - For ssh, only use port 22.
-- Merged from branch [update-ssh](https://github.com/shanejorr/usincometaxes/tree/update-ssh)
 
 # usincometaxes 0.3.0